stripe added

+++ hackathon1/src/app/api/cart/[id]/route.ts
@@ -3,6 +3,7 @@ import { eq } from "drizzle-orm";
 import { drizzle } from "drizzle-orm/vercel-postgres";
 import { NextRequest, NextResponse } from "next/server";
 import { pgTable, integer, text, serial } from "drizzle-orm/pg-core";
+import { auth } from "@clerk/nextjs";
 
 const cart = pgTable("cart", {
   _id: serial("_id").primaryKey(),
@@ -16,10 +17,15 @@ export const DELETE = async (
   { params }: { params: { id: string } }
 ) => {
   try {
-    console.log(params.id);
-
     const db = drizzle(sql);
+    const { userId } = auth();
+    if (!userId)
+      return NextResponse.json({
+        status: 401,
+        data: "You are unauthorized user.!",
+      });
     const resp = await db.delete(cart).where(eq(cart._id, Number(params.id)));
+    console.log("🚀 ~ file: route.ts:28 ~ resp:", resp);
     if (resp.rowCount) {
       return NextResponse.json({
         status: 200,
@@ -31,6 +37,7 @@ export const DELETE = async (
       data: "The product is already removed from your cart!",
     });
   } catch (error) {
+    console.log("🚀 ~ file: route.ts:40 ~ error:", error);
     return NextResponse.json({ status: 500, error: error });
   }
 };
